ww_mca.m4: fix parameter index
There was an incorrect m4 parameter index which was causing the wrong
variable assignment to be emitted into the configure script.  It
resulted in:

```sh
static_components="$all_components ...component_name..."
```

instead of:

```sh
all_components="$all_components ...component_name..."
```

Which led to all manner of chaos.  This commit uses the right m4
parameter index ($3, not $4), which restores the correct output to
the configure script.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>